+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 8.1
 - Add File Attached to Salaries in Salaries.php & Accounting/functions.inc.php
 - SQL BILLING_FEES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
 - Add File Attached to Fees in StudentFees.php & Student_Billing/functions.inc.php
+- Fix User Widgets for Advanced Search / Print Class Lists exports in GetStaffList.fnc.php, GetStuList.fnc.php & Search.inc.php
 
 Changes in 8.0.4
 ----------------